Update keywords
Four of the keywords included in this file are contextual. In particular,
```
least predicate
greatest predicate
least lemma
greatest lemma
```
are _keyword phrases_ in Dafny, but `least` and `greatest` are not keywords by themselves. Also, `to` and `downto` are keywords in the context of a `for` statement, but are not reserved elsewhere. Does this grammar have the power to make these distinctions? (If not, `least`, `greatest`, `to`, and `downto` are probably rare enough that it won’t bother too many users.)

Some operators are syntactically a prefix of others. For example, `<=` and `<==` and `<==>`. Do these need to be ordered in a particular way in this file?