fix: `dafny.dotnetExecutablePath` works to install custom source
If `dotnet` is installed
and placed in `dafny.dotnetExecutablePath`
custom source should still be able to be built.

To do this `checkSupportedDotnetVersion`
needs to be the primary way to detect `dotnet`.
This has the side effect of both not double installing from `brew`
but also not requireing an install if `dotnet` is already installed.

The `wget` call is also removed
since there are already helpers to perform this get and unzip.
They are updated to take a `thing` to facilitate logging.
There is also a bit of update to the `Z3` section
to account for this change.

The largest issue with this was
`Source/DafnyCore/DafnyCore.csproj`
has direct `dotnet` build targets.
This means that call to build will fail
if `dotnet` is not on the `PATH`.
This is unfortunate, but…
So we add it for that specific command.

Finnaly, speed up the `git clone`
to both checkout the tag and not get history.

Added a few things to `.gitignore`
and simplified the ReadMe’s use of `vsce`