Always attempt to find a binary distribution (#316)
This updates the installation process to always try to download a binary
distribution, using the current OS _and_ architecture, and only build
from source if the download fails. This should allow us to support other
architectures seamlessly in the future (and support the case where, for
some reason, we fail to upload a binary for a certain version/OS/arch
combination).