Feat: Added the test command (#420)
Fixes #417
This PR adds the "Test" command in the contextual Dafny menu.

A bit of context. It's not possible to put several Main methods in
included files, but it's definitely possible to put tests in each file
as needed. Currently, to run tests, one typically right-clicks "Dafny >
build with custom arguments", and then replace everything with "test".
There is not even a shortcut for that, although the "test" command is
fully supported in Dafny. Moreover, I am adding the `--no-verify` flag
by default both on tests and run because
1) it's in the context of a development environment
2) The minimum required for `dafny run` and `dafnyt test` is that the
program resolves and the compiler can compile it.
3) Verification of the entire file before testing is at best redundant
because it was already verified in the IDE, at worst it forces the user
to add a bunch of "assume false" in their code just because they haven't
prove some assertions they conjectured.

I think it's fair to say that the usual workflow of testing is
complementary to verification, even that testing often precedes
verification.
If users want still to run regular tests and run with verification, they
can use the trick of "build with custom arguments" and set up their
custom command, or just use the terminal.