Always show all verifiable symbols in the test panels (#428)
### Changes
- Show "preparing verification" instead of "waiting for free solver"
when all verification symbols are in a queued state. This is OK for
Dafny 4.2.0, since "preparing verification" could mean "waiting for free
solver". For the upcoming Dafny, this situation usually happens when
Dafny is performing translation to Boogie and resolution to Boogie, so
the message "preparing verification" is more correct.
- Always show all known verification symbols in the test panel, instead
of only those included in the current file. Together with project mode,
this enables easily getting an overview of all verifiables available in
the project.

### Testing
- Opened a customer project. Triggered verification for everything in
the project. Inspected the status bar, gutter icons and the testing
panel.