Fix: Dafny plugins working again. (#453)
The plugins option of the IDE was no longer working because we were
outputting the index. It might not matter that much now that we have
project files, but it's now fixed and I also deprecated the options to
add plugins this way since regular options work as much as project
files.

Fixes https://github.com/dafny-lang/dafny/issues/4759

I did not bother writing tests because it's a deprecated option anyway.