Fix binary copying to temporary folder on custom path
Currently, on Linux, the executable ends up not being copied due to not having one of the expected extensions.
This leads to unexpected behavior when specifying a path to a custom installation, where the specified binary ends up not being copied.
This becomes a problem on OSes and distributions, such as NixOS, where the extension's installer doesn't work properly.

This change makes it so that the specified binary is copied, no matter its file name.