copy native libraries when copying a custom dafny to tmpdir
Builds of Dafny downloaded from the GitHub Releases page include all
necessary support libraries, so they don't depend on an existing .NET
runtime. When we copy a custom dafny to a tmpdir, we need to copy these
libraries too, or the language server will fail to start.

Windows libraries, being .dlls, were probably already copied, but on
macOS and Linux, shared libraries end in .dylib and .so, respectively.